+++ src/application/delete-link.use-case.ts
@@ -1,6 +1,7 @@
 import { NotFoundError } from "@/resources/errors/app-error";
 
 import { type DeleteLinkDTO } from "./dto/link.dto";
+import { TripStatus } from "./dto/trip.dto";
 import { type DeleteLinkPort } from "./ports/delete-link.port";
 import { type LinkRepositoryPort } from "./ports/link.repository.port";
 
@@ -8,10 +9,20 @@ export class DeleteLinkUseCase implements DeleteLinkPort {
   constructor(private readonly linkRepository: LinkRepositoryPort) {}
 
   async execute(input: DeleteLinkDTO): Promise<void> {
-    const link = await this.linkRepository.findById(input.id, input.tripId);
+    const link = await this.linkRepository.findLinkById(input);
 
     if (!link) {
       throw new NotFoundError("Link não encontrado.");
+    }
+
+    if (!link.trip) {
+      throw new Error("Link não está associado a uma viagem.");
+    }
+
+    if (link.trip.status === TripStatus.CANCELLED) {
+      throw new Error(
+        "Não é possível deletar um link de uma viagem cancelada.",
+      );
     }
 
     await this.linkRepository.delete(input);
